Remove unused label feature and rename hamburger file
