+++ src/hamburger.ts
@@ -88,9 +88,6 @@ export function createHamburgerMenu() {
       action: "import_network_tray_directly_as_root",
       label: "import_network_tray_directly_as_root",
     },
-    { action: "manageLabels", label: "ラベル管理" },
-    { action: "exportLabels", label: "ラベルをエクスポート" },
-    { action: "importLabels", label: "ラベルをインポート" },
     { action: "editTitle", label: "ページタイトルを編集" },
     { action: "uploadAll", label: "Upload All" },
     { action: "downloadAll", label: "Download All" },
